_archive/archive_legacy_20251204/requirements.txt
# CBI-V15 Base Requirements (minimal, Python 3.12-safe)
# Add extras incrementally as the new plan hardens.

# Core data + IO
pandas>=2.2.3
numpy>=1.26.4
pyarrow>=16.1.0

# GCP
google-cloud-bigquery>=3.23.1
google-cloud-secret-manager>=2.25.0

# ML baseline
lightgbm>=4.5.0
scikit-learn>=1.5.2

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.2
